Add completion contributor for NGXS actions and state files
Added a completion contributor to support code completion for NGXS actions and state files. The enhancement includes completion options for action, action with payload, and selectors with or without metadata. Also added new icons for completion prompts with respect to light and dark themes. Updated the methods in NgxsActionsPsiFileFactory and NgxsStatePsiFileFactory to include additional parameters allowing a new line to be added during code generation. This was needed to prepare the code area for the new additions from the completion contributor.